+++ bin/propolis-server/src/lib/vm/guest_event.rs
